api-automation/requirements.txt
pytest==7.4.0
requests==2.31.0
python-dotenv==1.0.0
pytest-html==4.0.0
pytest-xdist==3.5.0